feat: move preferences from drawer to modal
