+++ example/src/Examples/CardExample.tsx
@@ -12,7 +12,7 @@ import {
 } from 'react-native-paper';
 
 import { isWeb } from '../../utils';
-import { PreferencesContext } from '../PreferencesContext';
+import { usePreferences } from '../Preferences/usePreferences';
 import ScreenWrapper from '../ScreenWrapper';
 
 type Mode = 'elevated' | 'outlined' | 'contained';
@@ -21,7 +21,7 @@ const CardExample = () => {
   const { colors } = useTheme();
   const [selectedMode, setSelectedMode] = React.useState<Mode>('elevated');
   const [isSelected, setIsSelected] = React.useState(false);
-  const preferences = React.useContext(PreferencesContext);
+  const { toggleTheme } = usePreferences();
 
   const modes: Mode[] = ['elevated', 'outlined', 'contained'];
 
@@ -186,13 +186,7 @@ const CardExample = () => {
             </Text>
           </Card.Content>
         </Card>
-        <Card
-          style={styles.card}
-          onPress={() => {
-            preferences?.toggleTheme();
-          }}
-          mode={selectedMode}
-        >
+        <Card style={styles.card} onPress={toggleTheme} mode={selectedMode}>
           <Card.Title
             title="Pressable Theme Change"
             left={(props) => <Avatar.Icon {...props} icon="format-paint" />}
