+++ example/src/Samples/MainSample.tsx
@@ -1,0 +1,3 @@
+export default function MainSample() {
+  return null;
+}
